+++ RELEASE_NOTES.md
@@ -2,8 +2,8 @@ Release Notes
 =============
 
 ## 1.6.31
+* AzureFirewall: Supports availability zones
 * WebApps/Functions: Add support for vnet integration
-
 
 ## 1.6.30
 * WebApps/Functions: Specify connection string types